add basis for the notification-system in backend
Notifications consist of a base type and can be inherited with specific Context that might be required to be displayed in a notification.

We will be using SSE to connect User frontend with the backend and use scheduled tasks in order to send updates in a respective cycle around events.