Implement RewardNotifications + badge display in backend + dtos
We are now able to receive RewardNotifications in the frontend, when the user has reached a new point threshold for their respective reward type.

In addition to that, there is a new endpoint under /api/rewards/badge?type=[]&threshold=[]  which will fetch a badge.png for the according type and level which will be sent to the frontend with lazy loading.

Cleaned up some files and added DTOs to reduce recursion in json response bodies.